% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_download_helpers.R
\name{download_plots}
\alias{download_plots}
\title{Export plots in editable formats.}
\usage{
download_plots(filepath, data_btn)
}
\arguments{
\item{filepath}{character indicating the filename.  This is created
via the downloadHandler in the parent function}
}
\description{
Export plots in editable formats.
}
